Extract MatchGameView from the match route (Spec 2 Phase 1 foundation)
Move the match-play page body into a re-hostable MatchGameView component; the
route becomes a thin wrapper. Pure refactor — no behavior change. The view reads
its own tripId (useParams) + gameId (?game=), so the same component can be
mounted by the route AND by the leaderboard's game panel (next commits), since
both live under /trips/[tripId]/ and both carry ?game=.
